Upgrade reCAPTCHA.AspNetCore from 2.2.5 to 3.0.10 with API updates
Agent-Logs-Url: https://github.com/surveysolutions/surveysolutions/sessions/2a095c5e-5f98-4190-97d1-debcc47d2a3d

+++ src/UI/Shared/WB.UI.Shared.Web.Core/Services/CaptchaConfig.cs
@@ -5,7 +5,7 @@
         public int TimespanInMinutesCaptchaWillBeShownAfterFailedLoginAttempt { get; set; } = 5;
         public int CountOfFailedLoginAttemptsBeforeCaptcha { get; set; } = 5;
         public CaptchaProviderType CaptchaType { get; set; } = CaptchaProviderType.None;
-        public decimal RecaptchaV3MinimumScore { get; set; } = 0.5m;
+        public double RecaptchaV3MinimumScore { get; set; } = 0.5;
     }
 
     public enum CaptchaProviderType
